move: get image source helper to common
