+++ apps/ui/components/reservation-unit/Images.tsx
@@ -5,7 +5,7 @@ import dynamic from "next/dynamic";
 import { breakpoints } from "common/src/common/style";
 import type { ReservationUnitImageNode } from "common/types/gql-types";
 import Carousel from "../Carousel";
-import { getImageSource } from "@/modules/util";
+import { getImageSource } from "common/src/helpers";
 
 const Modal = dynamic(() => import("../common/Modal"));
 type Props = {
